# SDCC experiments
This branch is dedicated to learning the Small Device C Compiler in
order to produce a Free firmware for the Truly Ergonomic keyboard.

We will start by writing a few small programs to acquire some general
feel about the compiler and its ways, then slowly expand our focus
to implement a USB HID device, and then to add specific keyboard
functionality.

Each new step will be described in the commit message. If you want to
understand how this branch evolved, you can read it on [Literate
GitHub][] (which is just an online repository browser with a focus on
commits).

[Literate GitHub]: http://yurivkhan.github.io/literate-github?yurivkhan/teck/commits/sdcc

(Of course, you can use any other repository browser if you wish, but
Literate GitHub renders Markdown in commit messages for you.)