Default state: Handling `GET_DESCRIPTOR` (cont)
What’s left is the `device_descriptor` constant. For now, let’s use
this. We’ll flesh it out some time later.

(I’m using a different Product ID so as not to interfere with my main
keyboard, which is, of course, another Truly Ergonomic. 030C is the
Product ID in production mode; when in ISP DFU mode, the device
identifies itself as 030B.)