Default state: Handling Standard device requests
In the Default state, we only handle a couple of Standard requests.
These are `GET_DESCRIPTOR` and `SET_ADDRESS`.

The latter has one quirk. In response to it, we must take the address
assigned to us and assign it to the `UADDR` register. But we must not do
that until after the whole control transfer is over. So we just stash
the new address into a variable, to be dealt with later, and send a
zero-length status response.