Default state: Receiving requests
When the host sends us anything, we will receive a USB interrupt with
the flag `URXD0`. I’m assuming that the MG84FL54’s FIFO is able to
buffer up to 64 bytes of data (which can accommodate the USB request
setup token just fine) and give us a USB interrupt with the `RXSETUP`
flag set in the Endpoint 0’s `RXSTAT` register.

Now we need to read it out of the FIFO. To read the received data, we
clear `EDOVW` in `RXSTAT` and then repeatedly read the `RXDAT` register.

We then acknowledge the receipt of the Setup token, handle the request,
and tell the USB subsystem that we’re done with this transfer. If
request handling fails, we stall the endpoint 0.