Address state: Manage remote wakeup feature
Add another global to track whether remote wakeup is enabled or
disabled. Initialize it to disabled on startup and on reset. In the
handler for `GET_STATUS` at the Device, return current value. Handle
`SET_FEATURE` and `CLEAR_FEATURE` to set and clear this flag.

We will use this variable when actually implementing suspend and resume.