Address state: Handle `GET_STATUS` for Endpoint 0
In the Address state, we have to handle `GET_STATUS` for the
device (which we did in the last commit) and for the control endpoint.
Add another case branch in the switch by recipient in `usb_request`,
repeat dispatch by request type, then request ID, then endpoint index.
For Endpoint 0, tell we’re not halted.