Configuration descriptor: Data structures
Last time, we got to the Address state, but it’s not yet the end of bus
enumeration. At the very least, we need to have a Configuration
descriptor and to communicate it to the host when asked.

Let’s first define the structure types and constants related to the
Configuration descriptor, straight from the USB and HID specs.

Note that the HID descriptor contains an array of HID class descriptor
headers. We are going to have just one such header for the Report
descriptor.