Configuration descriptor: Actual data
In the Device descriptor we said we have one Configuration. So the host
will next ask us for the descriptor of Configuration 1, by sending us a
Device `GET_DESCRIPTOR` request with `wValue` containing the
Configuration descriptor type in the high byte and 1 (the configuration
index) in the low byte.

In response to this request, we need to transmit a contiguous buffer
containing the configuration descriptor itself, then each of the
Interface descriptors with all of their Endpoint descriptors
interleaved. As a HID device, we are also expected to send with each
Interface descriptor its HID descriptor, right before the Endpoint
descriptors; then the host will ask us for the class-specific Report
descriptor.

For now, we will have a single Interface, with a single Endpoint. (Later
we will want another Interface, when dealing with media keys.)

Here’s the overall structure of our descriptors:

* Configuration
  * Interface 0
    * HID for Interface 0
    * Endpoint 1
  * (later) Interface 1
    * (later) HID for Interface 1
    * (later) Endpoint 2

In order to have all descriptors in a contiguous block, we define a
structure type containing all descriptors, and a constant of that type.