Configuration descriptor: Handling requests
Now that we’ve defined the descriptors, let’s write the code to return
them when requested.

The Configuration descriptor is returned through the same standard
`GET_DESCRIPTOR` request targeted at the device. The low byte of
`wValue` will contain the index of the configuration being requested.
Note how the _index_ of the configuration is not necessarily the same as
its _value_.

The HID descriptor and the HID Report descriptor are requested at the
Interface. (Recall that the HID spec says it should be the interface that
claims to conform to the HID class, not the whole device.) Otherwise,
it’s the same old `GET_DESCRIPTOR` request.