Configuration descriptor: Report descriptor
According to the HID spec, the report descriptor, unlike other
descriptors, is not a fixed-structure data record. Rather, it is a byte
sequence encoding the structure of reports. The report descriptor is
composed of items, with each item having a tag and optionally a value.

The easiest way to define such a structure would be as an array of
bytes, with actual byte values and their meaning in comments. The
reference implementation in the MG84FL54 SDK goes this way:

```c
code BYTE KB_HID_REPORT[]=                       //
  { 0x05,0x01,                                   // Usage Page (Generic Desktop Control)
    0x09,0x06,                                   // Usage (Keyboard)
    0xA1,0x01,                                   // Collection (Application)
    […]
```

While technically possible, it is very much like programming in machine
code. You don’t really program in it, you program in at least an
assembly language, and have it automatically translated to machine code.

What we want here is a domain-specific language, a means for us to
specify the items in human-readable form that would be translated into
suitable binary form.

The HID spec says an item may be short or long, depending on how much
data it carries. It also says there are no long items defined yet, so we
are left with short ones. A short item may have no argument, or a 1-, 2-
or 4-byte argument, and there are specific formats for each data size.

Since there is no easy way to determine at compile time which data size
each item wants, we will just specify this ourselves, by defining four
macros, one for each size. Each macro will expand to the appropriate
sequence of bytes.

Here we define the reference report descriptor for a HID Boot keyboard,
transcribed straight from the HID spec. The Boot keyboard protocol is
designed to allow BIOSes to work with USB keyboards without having to
parse report descriptors. All that a BIOS is expected to do is tell the
device “hey I’m a dumb BIOS, go easy on me” and the device will then
send reports according to this reference report descriptor. When a real
OS boots up, it will reset the device in normal mode, and the keyboard
may then enable all kinds of fancy features that need a different report
format, like NKRO (Boot only allows 8+6KRO, with 8 modifiers and up to 6
normal keys) or media keys.