Default state: Tracking states
We will track and visualize the states by lighting different
combinations of LEDs while progressing through each state.

Initially, we are in the Powered state (both Caps and Num LEDs off).
When we get the reset signal, we transition to Default state (Caps off,
Num on). The other two states are left unused for now.