Using Git to move around on this branch
To clone this repository and check out this branch:

```
$ git clone -b sdcc https://github.com/yurivkhan/teck.git
```

I suggest that you then use `gitk sdcc` to navigate around the branch.
For example, find this commit in the list, right-click it and select
“Create new branch”. Then type in a name for your reading branch. Then
right-click the newly created branch and choose “Check out this
branch”.

To move to a different commit, right-click it and select “Reset _name_
branch to here”. When prompted for the reset type, choose Hard reset.